Make `watch` run after `build`, not simultaneously.
This makes the `watch` have the `build` task as a dependency.  While
this would not be ideal for a large project, I think it should be OK for
most projects (since gulp is fast; this wouldn't work for me in grunt).

I imagine an end-user of this repo will have much customized the gulp
configuration anyway by the time their project is large enough for this
to be a problem.