Work around Font revving issue.
.ttf fonts have an embedded timestamp, which cause the contents
of the file to change ever-so-slightly. This was a problem for
file reving, which generates a hash based on the contents of the file.
This meant that even if source files had not changed, the hash would
change with every recompile.

The solution was to only hash svg font files, then append the
generate hash to the ttf, eot, and woff files.