don't fail silently when task-config has errors
when task-config.json in the project has errors (missing comma, single
quotes, etc), the build silently fails and uses the default. this is
undesirable. get-task-config can detect a "doesn't exist" error and use
the default only in that case. userland errors (in the project's task
config) should be exposed by failing nosily.