Remove try/catch in get-task-config
From @jbielick in #358
When task-config.json in the project has errors
(missing comma, single quotes, etc), the build silently
fails and uses the default. This is unexpected and resulted
in a bunch of files being written to the wrong (default)
directory. The clean task and others might be especially
harmful. In any event, get-task-config can detect a "doesn't
exist" error and use the default only in that case. userland
errors (in the project's task config) should be exposed by
failing nosily.

Resolves #358