The babel-preset-react-app plugin complains when either BABEL_ENV or NODE_ENV is not set to either 'development', 'test' or 'production'. This forces using `BABEL_ENV='development' yarn run blendid` incantation when integrating React.
To fix this, at top of webpack-multi-config, added an enforcment to setup
BABEL_ENV to something. It first defaults to itself so that if already set,
we do not change anything. Then, default to NODE_ENV if set and finally
will be set to webpack env variable if both BABEL_ENV and NODE_ENV are
not set.