Fixed problems when all asset or code tasks are disabled
When gulp-sequence is faced with an empty array, it seems that it will skip
all following tasks in the sequence. For example, facing a call like this:
gulp.sequence('clean', [], ['build', 'assets'], 'watch'), it will only
execute 'clean'. This is caused by the array being empty.

The problem is that if all assets tasks are disabled, it shortcuts the
overall sequence which prevents code tasks to be executed for example.

To overcome this, when building assetTasks and codeTasks objets, we now
set them to false if there is no tasks enabled.