Remove Karma + Mocha + Chai + Sinon setup (#416)
* Remove Karma + Mocha + Chai + Sinon setup

Just use Jest!

It used to be super complicated to string the right
series of tools together to get a cohesive and full
featured JS test suite — which is why we previously
did it for you. But now Jest exists, solves these
issues and is our strong recommendation.

This cuts out a TON of dependencies, and significantly
lightens up the package.

* Add chai back as devDependency