Prevent `browserSync.server.middleware` from being ovewritten completely
Currently, the `server.middleware` config option is completely set in stone
and there is no possibility to extend it. However, to be able to make React
Router `<BrowserRouter>` (opposed to `<HashRouter>`) to work, one need
to add a `browserSync` middleware.

To overcome the problem, two mecanisms has been added to solve this. First,
if the `browserSync.server.middleware` is already defined, blendid does not
try to override it anymore.

If not present, default middleware (`webpack-dev-middleware`) is used.
Furthermore, the default mdidlewares from `browserSync.server.middleware`
are concatenated with the ones find on config
`browserSync.server.extraMiddlewares` (defaults to `[]` when none defined).