Fix resolution of calling project `package.json` file (Fixes #532)
It appears that `process.env.PWD` is not consistent across platform as well
as across sub-directory invocations. Indeed, even if it is consistent on
Mac OS X (and probably Unix at the same time but untested), there is
inconsitencies problem on Windows when using a POSIX transformer platform
like Cygwin or MSYS2.

Indeed, in those cases, `PWD` is not the canonical Windows path but the
original POSIX path of the terminal. Furthermore, invoking from a
sub-directory also fiddles with the `PWD` adding subdirectory into it.

To fix that, it seems that Gulp is providing an environment variable that
points to the root directory of the project and which is constant across
platforms. See https://github.com/gulpjs/gulp/commit/2bf23ba407a594ac8ee7e693ec93b17dd906f926
for the actual INIT_CWD implementation.

So, to fix the problem everywhere, added a function `resolveRelativeToProject`
that does the correct resolving using `INIT_CWD` environment variable
instead of the previously used `PWD`.

To make stuff easier to use next time, a `projectPath` function was added
that does the resolving of paths relative to the project root directory.
All previous usages of `path.(resolve|join)(process.env.PWD, ...)` that were
used throughout the website are now using `projectPath(...)` instead.

Fixed bin script `blendid/gulpfile.js/index.js` resolution to use `__dirname`
instead of `require.resolve`. The `require.resolve` does not work when using
`yarn link` to development `blendid` features. Indeed, `require.resolve` does
the resolution from the linked project leading to `blendid` not being
resolvable (because not present in the linked `node_modules` directory). By
using `__dirname`, we achieve the same goal as before but being more robust
to linked project for development.