Added fs-extra as a direct dependency of Blendid
The fs-extra is required directly by Blendid in file
`gulpfile.js/tasks/replace-files.js`. However, this dependency was expected
to be present transitively via `browser-sync`.

Since it's a direct dependency of Blendid, it should be listed as such in
the `package.json`. By not being present, it was causing us problem when
using Yarn workspaces where `fs-extra` was not resolvable correctly (
because it was in different `node_modules` folder).

Used version `fs-extra@3.0.1` to remain the same as `browser-sync#fs-extra`.